Bot API 5.0 (November 4, 2020)

+++ src/Telegram.Bot/Requests/Available Methods/GetChatMemberRequest.cs
@@ -21,14 +21,14 @@ namespace Telegram.Bot.Requests
         /// Unique identifier of the target user
         /// </summary>
         [JsonProperty(Required = Required.Always)]
-        public int UserId { get; }
+        public long UserId { get; }
 
         /// <summary>
         /// Initializes a new request with chatId and userId
         /// </summary>
         /// <param name="chatId">Unique identifier for the target group or username of the target supergroup or channel</param>
         /// <param name="userId">Unique identifier of the target user</param>
-        public GetChatMemberRequest(ChatId chatId, int userId)
+        public GetChatMemberRequest(ChatId chatId, long userId)
             : base("getChatMember")
         {
             ChatId = chatId;
