+++ src/Telegram.Bot/Requests/Stickers/AddStickerToSetRequest.cs
@@ -18,7 +18,7 @@ namespace Telegram.Bot.Requests
         /// User identifier of sticker set owner
         /// </summary>
         [JsonProperty(Required = Required.Always)]
-        public int UserId { get; }
+        public long UserId { get; }
 
         /// <summary>
         /// Sticker set name
@@ -51,7 +51,7 @@ namespace Telegram.Bot.Requests
         /// <param name="name">Sticker set name</param>
         /// <param name="pngSticker">Png image with the sticker</param>
         /// <param name="emojis">One or more emoji corresponding to the sticker</param>
-        public AddStickerToSetRequest(int userId, string name, InputOnlineFile pngSticker, string emojis)
+        public AddStickerToSetRequest(long userId, string name, InputOnlineFile pngSticker, string emojis)
             : base("addStickerToSet")
         {
             UserId = userId;
